fix router with url change; scaffold routes
removed loadingFor and used window.location.hash
this already tells what it is at. and hash variable in the function
tells what route it has loaded. so we can check if the route has changed
while loading with this 2.

also loadingFor wasn't chaning when we change from url, so it's better.